clang tidy

+++ ecal/core/src/cimpl/ecal_client_cimpl.cpp
@@ -132,14 +132,14 @@ int eCAL_Client_AddResponseCallback(ECAL_HANDLE handle_, ResponseCallbackCT call
   if (handle_ == nullptr) return(0);
   auto* client = static_cast<eCAL::CServiceClient*>(handle_);
   auto callback = std::bind(g_response_callback, std::placeholders::_1, callback_, par_);
-  return client->AddResponseCallback(callback);
+  return static_cast<int>(client->AddResponseCallback(callback));
 }
 
 int eCAL_Client_RemResponseCallback(ECAL_HANDLE handle_)
 {
   if (handle_ == nullptr) return(0);
   auto* client = static_cast<eCAL::CServiceClient*>(handle_);
-  return client->RemResponseCallback();
+  return static_cast<int>(client->RemResponseCallback());
 }
 
 ECALC_API int eCAL_Client_AddEventCallback(ECAL_HANDLE handle_, enum eCAL_Client_Event type_, ClientEventCallbackCT callback_, void* par_)
